Dissolve SBML root element validation into sanity check.
