+++ validated-rules.txt
@@ -39,6 +39,8 @@
 10402
 10404
 10805
+20102
+20103
 20108
 20201
 20205
